requirements.txt
fastapi
uvicorn
requests
alpaca-py
pydantic
ratelimit
